+++ src/compat/maestro/export-flow.ts
@@ -1,4 +1,4 @@
-import type { SessionAction } from '../../daemon/types.ts';
+import type { SessionAction } from '../../contracts/session-action.ts';
 import { GESTURE_FLING_DURATION_MS } from '../../contracts/gesture-plan.ts';
 import { parseSelectorChain, type Selector } from '../../selectors/index.ts';
 import type { SelectorTerm } from '../../selectors/parse.ts';
