+++ src/daemon/daemon-command-registry.ts
@@ -7,23 +7,10 @@ export type { DaemonCommandRoute } from './request-handler-chain.ts';
 
 export type SessionCommandKind = 'inventory' | 'state' | 'observability' | 'publication' | 'replay';
 
-/**
- * ADR 0014 session ref-frame lifetime. Declares how a daemon command relates to
- * the session's authorized ref frame:
- * - `preserve`: no successful path changes device-visible element identity, so
- *   the frame carries through untouched (snapshots, reads, inventory, ...);
- * - `may-invalidate`: some successful path crosses a device side effect, so the
- *   leaf must expire the frame at its side-effect seam when that path runs;
- * - `delegated`: an orchestrator (batch/replay/test) whose nested leaves own
- *   their own transitions — the outer command never expires a frame itself.
- *
- * This classification is an honesty/completeness guard, NOT the transition site:
- * a `may-invalidate` command still calls the ref-frame module only when its
- * mutating path is selected. The completeness gate
- * (`__tests__/ref-frame-effect.test.ts`) fails if a daemon-projected command
- * omits this classification.
- */
-export type RefFrameEffect = 'preserve' | 'may-invalidate' | 'delegated';
+// Declared in contracts/ so core/ can classify commands without importing the daemon;
+// re-exported here because the descriptor shape below is stated in terms of it.
+export type { RefFrameEffect } from '../contracts/ref-frame-effect.ts';
+import type { RefFrameEffect } from '../contracts/ref-frame-effect.ts';
 
 /**
  * Request-sensitive form of {@link RefFrameEffect}. Commands whose subactions
